+++ src/GeneaLabs/Bones/Keeper/Entities/Validators/ModifyEntityValidator.php
@@ -1,9 +1,9 @@
 <?php namespace GeneaLabs\Bones\Keeper\Entities\Validators;
 
-use GeneaLabs\Bones\Keeper\BonesKeeperBaseValidator;
 use GeneaLabs\Bones\Keeper\Entities\Commands\ModifyEntityCommand;
+use GeneaLabs\Bones\Marshal\BonesMarshalBaseValidator;
 
-class ModifyEntityValidator extends BonesKeeperBaseValidator
+class ModifyEntityValidator extends BonesMarshalBaseValidator
 {
     protected static $rules = [
         'name' => 'required',